Remove the security group jazz
This jazz is being moved to cool-sharedservices-freeipa so that this
repo can be used for both IPA replicas and masters.  Once a replica is
up and running it is indistinguishable from a master.

Note that this means that a new variable, called security_group_ids is
now required.

Also update documentation and examples accordingly.