Have cloud-init create a file containing non-sensitive IPA config params
Also force the user to supply an IP address for the IPA server.  This
is done in order to break the dependency of DNS record resources on
EC2 instance resources; otherwise, it is impossible to replace the IPA
servers one by one as is required when a new AMI is generated.

Make realm an optional variable, since it is not needed for replicas.